load qTable from a file solved

+++ src/Weights.cpp
@@ -29,26 +29,12 @@ Weights::Weights(std::vector<float> w, Setting* configurationPtr)
 
 //constructor para qTableW
 Weights::Weights(std::vector< std::vector< std::vector<float> > > statePosibles, Setting* configurationPtr){
+	// struct fann *ann = fann_create_standard(4, 2, 8, 9, 1);
 	myType = (*configurationPtr)["type"].c_str();
 	gamma = (*configurationPtr)["relativeValue"];
-	/**************************************************************************************/
-	/*	Instanciacion del vector de posibles outputs / decisiones (combinaciones de pesos)*/
-	/*	sagun la discretizacion tomada de la configuracion								  */
-	/**************************************************************************************/
 	int wCantDiscretizacion = (*configurationPtr)["wPosibles"];
-	float step = 1.0 / (wCantDiscretizacion-1);
-	std::vector<float> wValPosibles;
-	for (int i = 0; i < wCantDiscretizacion; ++i)
-	{
-		wValPosibles.push_back(step * i);
-	}
-
-	//la suma de los pesos (en primera instancia) debe ser 1. La estrategia es hacer todas las combinaciones posibles de valores
-	//multiplos del step entre 0 y 1 (para step 0.2 [0.0 0.2 0.4 0.6 0.8 1.0]) y almacenar aquellas combinaciones donde la suma sea 1
-	std::vector<float> individuo;
-	wPermutaciones(wValPosibles, individuo, statePosibles.size(), &wCombinacionesPosibles);
-	cout << "combinaciones W obtenidas " << wCombinacionesPosibles.size() << endl;
-
+	comm = (*configurationPtr)["command"].c_str();
+	instanciarWcombinaciones(wCantDiscretizacion, statePosibles.size());
 	/*	posibles estados */
 	std::vector< std::vector<float> > sCombinacionesPosibles;
 	std::vector< std::vector<float> > sValPosibles;
@@ -59,6 +45,7 @@ Weights::Weights(std::vector< std::vector< std::vector<float> > > statePosibles,
 			sValPosibles.push_back(*itb);
 		}
 	}
+	std::vector<float> individuo;
 	sPermutaciones(sValPosibles, individuo, sValPosibles.size(), &sCombinacionesPosibles);
 	cout << "combinaciones S obtenidas " << sCombinacionesPosibles.size() << endl;
 	// printPerm(sCombinacionesPosibles);
@@ -87,7 +74,8 @@ Weights::Weights(std::vector< std::vector< std::vector<float> > > statePosibles,
 	//Carga de la qTable
 	int numberOfNodes = inputs.size();
 	int nodeSize = sizeof(std::map<std::vector<float> , qTableOutput>::value_type);
-	qTable.get_allocator().allocate( nodeSize * numberOfNodes );
+	allocateNb = nodeSize * numberOfNodes;
+	allocP = qTable.get_allocator().allocate( allocateNb );
 	for (std::vector< std::vector<float> >::iterator i = inputs.begin(); i != inputs.end(); ++i)
 	{
 		qTable[*i] = out;
@@ -114,9 +102,23 @@ Weights::Weights(std::vector< std::vector< std::vector<float> > > statePosibles,
 	}
 }
 
+//constructor para qTableW
+Weights::Weights(Setting* configurationPtr){
+	myType = (*configurationPtr)["type"].c_str();
+	int wCantDiscretizacion = (*configurationPtr)["wPosibles"];
+	int wSize = (*configurationPtr)["wSize"];
+	instanciarWcombinaciones(wCantDiscretizacion, wSize);
+	std::string file= (*configurationPtr)["file"].c_str();
+	int stateSize = (*configurationPtr)["sSize"];
+	int weightSize = (*configurationPtr)["wSize"];
+	int inputSize = stateSize + weightSize;
+	loadQTable(file,inputSize);
+	int aux = writeQTableToFile("./src/steering_behaviors_controller/qTableTest.txt");
+	cout << "Se almaceno en " << QTABLEFILE << " " << aux << " entradas." << endl;
+}
 
 Weights::~Weights(){
-
+	qTable.get_allocator().deallocate(allocP,allocateNb);
 }
 
 std::vector<float> Weights::getWeights(std::vector< std::vector<float> > state){
@@ -127,9 +129,9 @@ std::vector<float> Weights::getWeights(std::vector< std::vector<float> > state){
 			return updateConstW(state);
 		}
 		return *weights;
-	}else if (myType == "qvalueW")
-	{
+	} else {
 		std::vector<float> statePart;
+		std::vector<float> wAux;
 		for (std::vector< std::vector< float > >::iterator ita = state.begin(); ita != state.end(); ++ita)
 		{
 			for (std::vector<float>::iterator itb = ita->begin(); itb != ita->end(); ++itb)
@@ -137,27 +139,30 @@ std::vector<float> Weights::getWeights(std::vector< std::vector<float> > state){
 				statePart.push_back(*itb);
 			}
 		}
-		std::vector<float> wAux = getWfromQTable(statePart);
-
-		//Verifico que el estado no corresponde a ningun refuerzo
-		int testigo = criticCheck(state);
-		if (testigo != -1) {
-			//si corresponde a algun refuerzo, se actualizan los valores de la tabla,
-			actualizarQTable(testigo);
-			//se limpia la memoria,
-			memoria.clear();
-			//se reinicia la simulacion
-			system("killall stageros &");
-			sleep(1);
-			system("rosrun stage_ros stageros /home/nahuel/catkin_ws/src/steering_behaviors_controller/world/willow-four-erratics.world &");
-			sleep(1);
-			//y se devuelve un vector de pesos nulos, para reiniciar aprendizaje a partir del proximo estado
-			std::vector<float> wNull;
-			for (int i = 0; i < state.size(); ++i)
-			{
-				wNull.push_back(0.000);
+		if (myType == "qvalueW") {
+			wAux = getRandomWfromQTable(statePart);
+			//Verifico que el estado no corresponde a ningun refuerzo
+			int testigo = criticCheck(state);
+			if (testigo != -1) {
+				//si corresponde a algun refuerzo, se actualizan los valores de la tabla,
+				actualizarQTable(testigo);
+				//se limpia la memoria,
+				memoria.clear();
+				//se reinicia la simulacion
+				system("killall stageros &");
+				sleep(1);
+				system(comm.c_str());
+				sleep(1);
+				//y se devuelve un vector de pesos nulos, para reiniciar aprendizaje a partir del proximo estado
+				std::vector<float> wNull;
+				for (int i = 0; i < state.size(); ++i)
+				{
+					wNull.push_back(0.000);
+				}
+				wAux = wNull;
 			}
-			wAux = wNull;
+		} else if (myType == "constQvalueW") {
+			wAux = getBestWfromQTable(statePart);
 		}
 		return wAux;
 	}
@@ -268,7 +273,30 @@ int Weights::writeQTableToFile(std::string fname) {
 	return count;
 }
 
-std::vector<float> Weights::getWfromQTable(std::vector<float> state){
+std::vector<float> Weights::getRandomWfromQTable(std::vector<float> state){
+	//Genero todos los posibles inputs estado/pesos, correspondientes al estado actual
+	std::vector< std::vector<float> > options;
+	for (std::vector< std::vector<float> >::iterator itw = wCombinacionesPosibles.begin(); itw != wCombinacionesPosibles.end(); ++itw)
+	{
+		std::vector<float> aux = state;
+		aux.insert( aux.end(), itw->begin(), itw->end() );
+		options.push_back(aux);
+	}
+	//Extraigo de la qTable los valores de visitas y qValue correspondiente a las posibilidades
+	std::vector<qTableOutput> outputs;
+	for (std::vector< std::vector<float> >::iterator iti = options.begin(); iti != options.end(); ++iti)
+	{
+		outputs.push_back(qTable[*iti]);
+	}
+	//Eleccion aleatoria
+	int aleatorio = rand()% outputs.size();
+	//Agrego el mejor elegido a la lista de estados visitados, para la posterior actualizacion de los qValues correspondientes en funcion de los refuerzos recibidos en un futuro
+	std::map<std::vector<float> , qTableOutput>::iterator itaux = qTable.find(options[aleatorio]);
+	memoria.push_back(itaux);
+	return wCombinacionesPosibles[aleatorio];
+}
+
+std::vector<float> Weights::getBestWfromQTable(std::vector<float> state){
 	//Genero todos los posibles inputs estado/pesos, correspondientes al estado actual
 	std::vector< std::vector<float> > options;
 	for (std::vector< std::vector<float> >::iterator itw = wCombinacionesPosibles.begin(); itw != wCombinacionesPosibles.end(); ++itw)
@@ -284,37 +312,18 @@ std::vector<float> Weights::getWfromQTable(std::vector<float> state){
 		outputs.push_back(qTable[*iti]);
 	}
 	//Evaluando las salidas elijo la mejor opcion para el estado actual
-	// int best = 0;
-	// int index = 0;
-	// for (std::vector<qTableOutput>::iterator ito = outputs.begin(); ito != outputs.end(); ++ito, index++)
-	// {
-	// 	if ((ito->qValue)>(outputs[best].qValue))
-	// 	{
-	// 		best = index;
-	// 	}
-	// }
-	//Eleccion aleatoria
-	int best = rand()% outputs.size();
-	//Agrego el mejor elegido a la lista de estados visitados, para la posterior actualizacion de los qValues correspondientes en funcion de los refuerzos recibidos en un futuro
-	std::map<std::vector<float> , qTableOutput>::iterator itaux = qTable.find(options[best]);
-	memoria.push_back(itaux);
-
-	//SOLO PARA TEST
-	// cout << "Se almaceno en la memoria puntero al elemento de la qTable ";
-	// std::vector<float> auxv = (memoria.back())->first;
-	// for (std::vector<float>::iterator iv = auxv.begin(); iv != auxv.end(); ++iv)
-	// {
-	// 	cout << *iv << " ";
-	// }
-	// qTableOutput auxo = (memoria.back())->second;
-	// cout << "= " << auxo.visits << " " << auxo.qValue << endl;
-	// cout << "getWfromQTable returns ";
-	// for (std::vector<float>::iterator i = wCombinacionesPosibles[best].begin(); i != wCombinacionesPosibles[best].end(); ++i)
-	// {
-	// 	cout << *i << " ";
-	// }
-	// cout << endl;
-
+	int best = 0;
+	int index = 0;
+	float bestQval=outputs[best].qValue;
+	for (std::vector<qTableOutput>::iterator ito = outputs.begin(); ito != outputs.end(); ++ito, index++)
+	{
+		if ((ito->qValue)>(outputs[best].qValue))
+		{
+			best = index;
+			bestQval =ito->qValue;
+		}
+	}
+	cout<< "lo mejor que habia " << bestQval<< endl;
 	return wCombinacionesPosibles[best];
 }
 
@@ -337,53 +346,92 @@ int Weights::criticCheck(std::vector< std::vector<float> > state){
 	return -1;
 }
 
-// void Weights::actualizarQTable(int refuerzo){
-// 	std::string mensaje = critic[refuerzo].message;
-// 	cout << "Aplicando refuerzo " << mensaje << " a " << memoria.size() << " estados" << endl;
-// 	//a cada elemento del mapa en la memoria hacer visitas++ y cambiar el qval segun lo q diga el refuerzo
-// 	int index = 0;
-// 	for (std::vector< std::map<std::vector<float> , qTableOutput>::iterator >::iterator itmem = memoria.end(); itmem != memoria.begin(); --itmem, index++)
-// 	{
-// 		std::vector<float> printv = (*itmem)->first;
-// 		cout << "actualizando " << index << "estado: ";
-// 		for (std::vector<float>::iterator itv = printv.begin(); itv < printv.end(); itv++) {
-// 			cout << *itv << " ";
-// 		}
-// 		if (itmem == memoria.end()) {
-// 			(*itmem)->second.qValue = critic[refuerzo].reinforcementValue;
-// 		} else {
-// 			(*itmem)->second.qValue += gamma*(*(itmem+1))->second.qValue;
-// 		}
-// 		(*itmem)->second.visits++;
-// 		cout << "qVal = " << (*itmem)->second.qValue << endl;
-// 	}
-// 	memoria.clear();
-// 	int aux = writeQTableToFile(QTABLEFILE);
-// 	cout << "Se almaceno en " << QTABLEFILE << " " << aux << " entradas." << endl;
-// }
-
 void Weights::actualizarQTable(int refuerzo){
 	std::string mensaje = critic[refuerzo].message;
 	cout << "Aplicando refuerzo " << mensaje << " a " << memoria.size() << " estados" << endl;
 
 	//Al ultimo estado le asigno el valor de la recompensa
-	(*(memoria.back())).second.qValue += critic[refuerzo].reinforcementValue;
+	(*(memoria.back())).second.qValue = critic[refuerzo].reinforcementValue;
 	(*(memoria.back())).second.visits++;
+	cout << endl;
 	//los estados anteriores se actualizan en funcion de este refuerzo recibido
-	//a cada elemento del mapa en la memoria hacer visitas++ y cambiar el qval segun lo q diga el refuerzo
-	int index = 0;
-	for (std::vector< std::map<std::vector<float> , qTableOutput>::iterator >::iterator itmem = memoria.begin(); itmem != (memoria.end()-1); ++itmem, index++)
-	{
-		std::vector<float> printv = (*itmem)->first;
-		cout << "actualizando " << index << " estado: ";
-		for (std::vector<float>::iterator itv = printv.begin(); itv < printv.end(); itv++) {
-			cout << *itv << " ";
+	std::vector<float> printv = (*(memoria.back())).first;
+	for (std::vector<float>::iterator itv = printv.begin(); itv < printv.end(); itv++){
+		cout << std::fixed << std::setprecision(3) << *itv << " ";
+	}
+	cout << "\t REFUERZO = " << (*(memoria.back())).second.qValue << endl;
+	if (memoria.size() > 1) {
+		for (std::vector< std::map<std::vector<float> , qTableOutput>::iterator >::iterator itmem = (memoria.end()-1) ; itmem != memoria.begin(); --itmem)
+		{
+			std::vector<float> printv = (*itmem)->first;
+			for (std::vector<float>::iterator itv = printv.begin(); itv < printv.end(); itv++){
+				cout << std::fixed << std::setprecision(3) << *itv << " ";
+			}
+			cout << "\t qValProm = " << (*itmem)->second.qValue << " qvalAdd= " << gamma << " * " << (*(itmem+1))->second.qValue;
+			//Como la sucesión de estados no es estocastica, se promedia el qvalue
+			(*itmem)->second.qValue = (((*itmem)->second.qValue * (*itmem)->second.visits) +gamma*(*(itmem+1))->second.qValue)/((*itmem)->second.visits + 1);
+			(*itmem)->second.visits++;
+			cout << " -> " << (*itmem)->second.qValue << endl;
 		}
-		(*itmem)->second.qValue += gamma*(*(itmem+1))->second.qValue;
-		(*itmem)->second.visits++;
-		cout << "qVal = " << (*itmem)->second.qValue << endl;
 	}
 	memoria.clear();
 	int aux = writeQTableToFile(QTABLEFILE);
 	cout << "Se almaceno en " << QTABLEFILE << " " << aux << " entradas." << endl;
+}
+
+void Weights::loadQTable(std::string file, int inputSize){
+	std::ifstream aFile (file.c_str());
+	int numberOfNodes=std::count(std::istreambuf_iterator<char>(aFile), std::istreambuf_iterator<char>(), '\n');
+	int nodeSize = sizeof(std::map<std::vector<float> , qTableOutput>::value_type);
+	allocateNb = nodeSize * numberOfNodes;
+	allocP = qTable.get_allocator().allocate( allocateNb );
+	cout << allocateNb << "=" << nodeSize << "*" << numberOfNodes << endl;
+	//para cada linea
+	aFile.seekg(0);
+	for (size_t line = 0; line < numberOfNodes; line++) {
+		//extraigo el input
+		std::vector<float> inV;
+		for (size_t in = 0; in < inputSize; in++) {
+			float aux;
+			aFile >> aux;
+			inV.push_back(aux);
+		}
+		//extraigo el Output
+		char dummy; //para el "="
+		qTableOutput outAux;
+		aFile >> dummy >> outAux.visits >> outAux.qValue ;
+		//lo cargo al mapa
+		qTable[inV] = outAux;
+	}
+	cout << "Se copiaron a la tabla final " << qTable.size() << endl;
+	printQTable();
+}
+
+void Weights::instanciarWcombinaciones(int wCantDiscretizacion, int size){
+	/**************************************************************************************/
+	/*	Instanciacion del vector de posibles outputs / decisiones (combinaciones de pesos)*/
+	/*	sagun la discretizacion tomada de la configuracion								  */
+	/**************************************************************************************/
+	float step = 1.0 / (wCantDiscretizacion-1);
+	std::vector<float> wValPosibles;
+	for (int i = 0; i < wCantDiscretizacion; ++i)
+	{
+		wValPosibles.push_back(step * i);
+	}
+	//la suma de los pesos (en primera instancia) debe ser 1. La estrategia es hacer todas las combinaciones posibles de valores
+	//multiplos del step entre 0 y 1 (para step 0.2 [0.0 0.2 0.4 0.6 0.8 1.0]) y almacenar aquellas combinaciones donde la suma sea 1
+	std::vector<float> individuo;
+	wPermutaciones(wValPosibles, individuo, size, &wCombinacionesPosibles);
+	cout << "combinaciones W obtenidas " << wCombinacionesPosibles.size() << endl;
+}
+
+void Weights::printQTable(){
+	for(std::map<std::vector<float> , qTableOutput>::iterator itm = qTable.begin(); itm != qTable.end(); itm++) {
+		std::vector<float> auxv = itm->first;
+		for (std::vector<float>::iterator itv = auxv.begin(); itv != auxv.end(); ++itv)
+		{
+			cout << (*itv) << " ";
+		}
+		cout << "=" << itm->second.visits << " " << itm->second.qValue << endl;
+	}
 }
